# perf: reuse the same buffer in unit_propagation