Using `dhat-rs` I saw that most of our allocations are short lived allocation inlined into `unit_propagation`. The only allocation in there is the calls to `vec![`, so this saves one vec in State so we can reuse the same buffer.

Running the benchmarks this provides a major speed up when applied on top of #73

```
Warning: Unable to complete 100 samples in 20.0s. You may wish to increase target time to 30.3s, or reduce sample count to 60.
large_cases/elm-packages_str_SemanticVersion.ron                                                                            
                        time:   [301.46 ms 301.66 ms 301.86 ms]
                        change: [-19.093% -19.010% -18.927%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild
large_cases/large_case_u16_NumberVersion.ron                                                                             
                        time:   [16.784 ms 16.798 ms 16.813 ms]
                        change: [-21.591% -21.498% -21.406%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  8 (8.00%) high mild
  1 (1.00%) high severe

```