@mpizenberg 2021-01-23
To me this does not provide a significant speedup. I've merged #73 in dev. No comparing this to dev, I obtain a 2% regression on the large case benchmark and a 2% improvement on the benchmark for elm packages. Maybe the difference is due to different system allocators?

I like the `while let ...` though :)