@mpizenberg 2021-01-23
By the way, I'm always stashing and unstashing things for the benchmark to run, could you @Eh2406 make PR with your current benchmarking code able to handle different types of `.ron` files in `test-examples/`?