@Eh2406 2021-01-23
> Maybe the difference is due to different system allocators?

I have been testing on win 10. I can try to run it on a linux box for comparison. It makes me wonder if some of the other changes that are perf neutral on my set up my be big on someone else's.