@Eh2406 2021-01-24
15% and 20% improvement on my linux boot with the default allocator. 12% and 24% improvement with jemalloc.
In what configuration are you running?