@mpizenberg 2021-01-24
I'm on archlinux, rustc 1.49, default allocator. I'll retry to make sure I didn't make a mistake.